refactor wip: move explorer mock routes to mockSiteRouter
